backend/requirements.txt
# Django
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.6.0

# Authentication
djangorestframework-simplejwt==5.4.0
PyJWT==2.10.1

# Database
psycopg2-binary==2.9.10

# Async / Channels
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
uvicorn[standard]==0.34.0

# Celery
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.1

# Biometric / Face Recognition
face_recognition==1.3.0
dlib==19.24.6
Pillow==11.1.0
numpy==1.26.4

# Encryption
cryptography==44.0.0

# API
drf-spectacular==0.28.0

# Utilities
python-decouple==3.8
django-extensions==3.2.3
whitenoise==6.8.2

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-asyncio==0.24.0
factory-boy==3.3.1
faker==33.1.0

# Production
gunicorn==23.0.0

# Requests (for webhooks)
requests==2.32.3
httpx==0.28.1
